+++ frontend/src/components/AppLayout/SearchBar.module.css
